+++ src/main/clojure/dev/eca/eca_intellij/extension/init_server_startup.clj
@@ -3,7 +3,6 @@
    [com.github.ericdallo.clj4intellij.extension :refer [def-extension]]
    [dev.eca.eca-intellij.db :as db]
    [dev.eca.eca-intellij.extension.server-logs :as server-logs]
-   [dev.eca.eca-intellij.listener.file :as listener.file]
    [dev.eca.eca-intellij.server :as server])
   (:import
    [com.intellij.openapi.project Project]
@@ -18,5 +17,4 @@
     (when-not (contains? #{:connected :connecting}
                          (server/status project))
       (server/start! project))
-    (listener.file/track-files! project)
     (db/assoc-in project [:on-stderr-log-updated-fns :editor-log] (fn [] (server-logs/update-logs! project)))))